refactor sidebar into separate module
