+++ sshpilot/window.py
@@ -38,6 +38,7 @@ from .askpass_utils import ensure_askpass_script
 from .preferences import PreferencesWindow, is_running_in_flatpak
 from .sshcopyid_window import SshCopyIdWindow
 from .groups import GroupManager
+from .sidebar import GroupRow, ConnectionRow, build_sidebar
 
 from .sftp_utils import open_remote_in_file_manager
 
@@ -45,395 +46,6 @@ logger = logging.getLogger(__name__)
 
 
 
-# ============================================================================
-# Advanced Grouping System
-# ============================================================================
-
-class GroupRow(Gtk.ListBoxRow):
-    """Row widget for group headers"""
-    
-    __gsignals__ = {
-        'group-toggled': (GObject.SignalFlags.RUN_FIRST, None, (str, bool))
-    }
-    
-    def __init__(self, group_info: Dict, group_manager: GroupManager, connections_dict: Dict = None):
-        super().__init__()
-        self.group_info = group_info
-        self.group_manager = group_manager
-        self.group_id = group_info['id']
-        self.connections_dict = connections_dict or {}
-        
-        # Create main content box
-        content = Gtk.Box(orientation=Gtk.Orientation.HORIZONTAL, spacing=12)
-        content.set_margin_start(12)
-        content.set_margin_end(12)
-        content.set_margin_top(8)
-        content.set_margin_bottom(8)
-        
-        # Group icon
-        icon = Gtk.Image.new_from_icon_name('folder-symbolic')
-        icon.set_icon_size(Gtk.IconSize.NORMAL)
-        content.append(icon)
-        
-        # Group info
-        info_box = Gtk.Box(orientation=Gtk.Orientation.VERTICAL, spacing=2)
-        info_box.set_hexpand(True)
-        
-        # Group name label
-        self.name_label = Gtk.Label()
-        self.name_label.set_markup(f"<b>{group_info['name']}</b>")
-        self.name_label.set_halign(Gtk.Align.START)
-        info_box.append(self.name_label)
-        
-        # Connection count label
-        self.count_label = Gtk.Label()
-        self.count_label.add_css_class('dim-label')
-        self.count_label.set_halign(Gtk.Align.START)
-        info_box.append(self.count_label)
-        
-        content.append(info_box)
-        
-        # Expand/collapse button
-        self.expand_button = Gtk.Button()
-        self.expand_button.set_icon_name('pan-end-symbolic')
-        self.expand_button.add_css_class('flat')
-        self.expand_button.add_css_class('group-expand-button')
-        self.expand_button.set_can_focus(False)
-        self.expand_button.connect('clicked', self._on_expand_clicked)
-        content.append(self.expand_button)
-        
-        self.set_child(content)
-        self.set_selectable(True)
-        self.set_can_focus(True)
-        
-        # Update display
-        self._update_display()
-        
-        # Setup drag source
-        self._setup_drag_source()
-        
-        # Setup double-click gesture for expand/collapse
-        self._setup_double_click_gesture()
-    
-    def _update_display(self):
-        """Update the display based on current group state"""
-        # Update expand/collapse icon
-        if self.group_info.get('expanded', True):
-            self.expand_button.set_icon_name('pan-down-symbolic')
-        else:
-            self.expand_button.set_icon_name('pan-end-symbolic')
-        
-        # Update connection count - only count connections that actually exist
-        actual_connections = []
-        for conn_nickname in self.group_info.get('connections', []):
-            if conn_nickname in self.connections_dict:
-                actual_connections.append(conn_nickname)
-        
-        count = len(actual_connections)
-        if count == 1:
-            self.count_label.set_text("1")
-        else:
-            self.count_label.set_text(f"{count}")
-    
-    def _on_expand_clicked(self, button):
-        """Handle expand/collapse button click"""
-        self._toggle_expand()
-    
-    def _setup_drag_source(self):
-        """Setup drag source for group reordering"""
-        drag_source = Gtk.DragSource()
-        drag_source.set_actions(Gdk.DragAction.MOVE)
-        drag_source.connect('prepare', self._on_drag_prepare)
-        self.add_controller(drag_source)
-    
-    def _on_drag_prepare(self, source, x, y):
-        """Prepare drag data"""
-        data = {
-            'type': 'group',
-            'group_id': self.group_id
-        }
-        return Gdk.ContentProvider.new_for_value(GObject.Value(GObject.TYPE_PYOBJECT, data))
-    
-    def _setup_double_click_gesture(self):
-        """Setup double-click gesture for expand/collapse"""
-        gesture = Gtk.GestureClick()
-        gesture.set_button(1)  # Left mouse button
-        gesture.connect('pressed', self._on_double_click)
-        self.add_controller(gesture)
-    
-    def _on_double_click(self, gesture, n_press, x, y):
-        """Handle double-click to expand/collapse group"""
-        if n_press == 2:  # Double-click
-            self._toggle_expand()
-    
-    def _toggle_expand(self):
-        """Toggle the expanded state of the group"""
-        expanded = not self.group_info.get('expanded', True)
-        self.group_info['expanded'] = expanded
-        self.group_manager.set_group_expanded(self.group_id, expanded)
-        self._update_display()
-        
-        # Emit signal to parent to rebuild the list
-        self.emit('group-toggled', self.group_id, expanded)
-
-
-class ConnectionRow(Gtk.ListBoxRow):
-    """Row widget for connection list"""
-    
-    def __init__(self, connection: Connection):
-        super().__init__()
-        self.connection = connection
-        
-        # Create overlay for pulse effect
-        overlay = Gtk.Overlay()
-        self.set_child(overlay)
-        
-        # Create main content box
-        content = Gtk.Box(orientation=Gtk.Orientation.HORIZONTAL, spacing=12)
-        content.set_margin_start(12)
-        content.set_margin_end(12)
-        content.set_margin_top(6)
-        content.set_margin_bottom(6)
-        
-        # Connection icon
-        icon = Gtk.Image.new_from_icon_name('computer-symbolic')
-        icon.set_icon_size(Gtk.IconSize.NORMAL)
-        content.append(icon)
-        
-        # Connection info
-        info_box = Gtk.Box(orientation=Gtk.Orientation.VERTICAL, spacing=2)
-        info_box.set_hexpand(True)
-        
-        # Nickname label
-        self.nickname_label = Gtk.Label()
-        self.nickname_label.set_markup(f"<b>{connection.nickname}</b>")
-        self.nickname_label.set_halign(Gtk.Align.START)
-        info_box.append(self.nickname_label)
-        
-        # Host info label (may be hidden based on user setting)
-        self.host_label = Gtk.Label()
-        self.host_label.set_halign(Gtk.Align.START)
-        self.host_label.add_css_class('dim-label')
-        self._apply_host_label_text()
-        info_box.append(self.host_label)
-        
-        content.append(info_box)
-        
-        # Port forwarding indicators (L/R/D)
-        self.indicator_box = Gtk.Box(orientation=Gtk.Orientation.HORIZONTAL, spacing=6)
-        self.indicator_box.set_halign(Gtk.Align.CENTER)
-        self.indicator_box.set_valign(Gtk.Align.CENTER)
-        content.append(self.indicator_box)
-
-        # Connection status indicator
-        self.status_icon = Gtk.Image.new_from_icon_name('network-offline-symbolic')
-        self.status_icon.set_pixel_size(16)  # GTK4 uses pixel size instead of IconSize
-        content.append(self.status_icon)
-        
-        # Set content as the main child of overlay
-        overlay.set_child(content)
-        
-        # Create pulse layer
-        self._pulse = Gtk.Box()
-        self._pulse.add_css_class("pulse-highlight")
-        self._pulse.set_can_target(False)  # Don't intercept mouse events
-        self._pulse.set_hexpand(True)
-        self._pulse.set_vexpand(True)
-        overlay.add_overlay(self._pulse)
-        
-        self.set_selectable(True)  # Make the row selectable for keyboard navigation
-        
-        # Update status
-        self.update_status()
-        # Update forwarding indicators
-        self._update_forwarding_indicators()
-        
-        # Setup drag source for connection reordering
-        self._setup_drag_source()
-    
-    def _setup_drag_source(self):
-        """Setup drag source for connection reordering"""
-        drag_source = Gtk.DragSource()
-        drag_source.set_actions(Gdk.DragAction.MOVE)
-        drag_source.connect('prepare', self._on_drag_prepare)
-        drag_source.connect('drag-begin', self._on_drag_begin)
-        drag_source.connect('drag-end', self._on_drag_end)
-        self.add_controller(drag_source)
-    
-    def _on_drag_prepare(self, source, x, y):
-        """Prepare drag data"""
-        data = {
-            'type': 'connection',
-            'connection_nickname': self.connection.nickname
-        }
-        return Gdk.ContentProvider.new_for_value(GObject.Value(GObject.TYPE_PYOBJECT, data))
-    
-    def _on_drag_begin(self, source, drag):
-        """Handle drag begin - show ungrouped area"""
-        try:
-            # Get the main window to show ungrouped area
-            window = self.get_root()
-            if hasattr(window, '_show_ungrouped_area'):
-                window._show_ungrouped_area()
-        except Exception as e:
-            logger.error(f"Error in drag begin: {e}")
-    
-    def _on_drag_end(self, source, drag, delete_data):
-        """Handle drag end - hide ungrouped area"""
-        try:
-            # Get the main window to hide ungrouped area
-            window = self.get_root()
-            if hasattr(window, '_hide_ungrouped_area'):
-                window._hide_ungrouped_area()
-        except Exception as e:
-            logger.error(f"Error in drag end: {e}")
-
-    @staticmethod
-    def _install_pf_css():
-        try:
-            # Install CSS for port forwarding indicator badges once per display
-            display = Gdk.Display.get_default()
-            if not display:
-                return
-            # Use an attribute on the display to avoid re-adding provider
-            if getattr(display, '_pf_css_installed', False):
-                return
-            provider = Gtk.CssProvider()
-            css = """
-            .pf-indicator { /* kept for legacy, not used by circled glyphs */ }
-            .pf-local { color: #E01B24; }
-            .pf-remote { color: #2EC27E; }
-            .pf-dynamic { color: #3584E4; }
-            """
-            provider.load_from_data(css.encode('utf-8'))
-            Gtk.StyleContext.add_provider_for_display(display, provider, Gtk.STYLE_PROVIDER_PRIORITY_APPLICATION)
-            setattr(display, '_pf_css_installed', True)
-        except Exception:
-            pass
-
-
-
-    def _update_forwarding_indicators(self):
-        # Ensure CSS exists
-        self._install_pf_css()
-        # Clear previous indicators
-        try:
-            while self.indicator_box.get_first_child():
-                self.indicator_box.remove(self.indicator_box.get_first_child())
-        except Exception:
-            return
-
-        rules = getattr(self.connection, 'forwarding_rules', []) or []
-        has_local = any(r.get('enabled', True) and r.get('type') == 'local' for r in rules)
-        has_remote = any(r.get('enabled', True) and r.get('type') == 'remote' for r in rules)
-        has_dynamic = any(r.get('enabled', True) and r.get('type') == 'dynamic' for r in rules)
-
-        def make_badge(letter: str, cls: str):
-            # Use Unicode precomposed circled letters for perfect centering
-            circled_map = {
-                'L': '\u24C1',  # Ⓛ
-                'R': '\u24C7',  # Ⓡ
-                'D': '\u24B9',  # Ⓓ
-            }
-            glyph = circled_map.get(letter, letter)
-            lbl = Gtk.Label(label=glyph)
-            lbl.add_css_class(cls)
-            lbl.set_halign(Gtk.Align.CENTER)
-            lbl.set_valign(Gtk.Align.CENTER)
-            try:
-                lbl.set_xalign(0.5)
-                lbl.set_yalign(0.5)
-            except Exception:
-                pass
-            return lbl
-
-        if has_local:
-            self.indicator_box.append(make_badge('L', 'pf-local'))
-        if has_remote:
-            self.indicator_box.append(make_badge('R', 'pf-remote'))
-        if has_dynamic:
-            self.indicator_box.append(make_badge('D', 'pf-dynamic'))
-
-    def _apply_host_label_text(self):
-        try:
-            window = self.get_root()
-            hide = bool(getattr(window, '_hide_hosts', False)) if window else False
-        except Exception:
-            hide = False
-        if hide:
-            self.host_label.set_text('••••••••••')
-        else:
-            self.host_label.set_text(f"{self.connection.username}@{self.connection.host}")
-
-    def apply_hide_hosts(self, hide: bool):
-        """Called by window when hide/show toggles."""
-        self._apply_host_label_text()
-
-    def update_status(self):
-        """Update connection status display"""
-        try:
-            # Check if there's any active terminal for this connection
-            window = self.get_root()
-            has_active_terminal = False
-
-            # Prefer multi-tab map if present; fallback to most-recent mapping
-            if hasattr(window, 'connection_to_terminals') and self.connection in getattr(window, 'connection_to_terminals', {}):
-                for t in window.connection_to_terminals.get(self.connection, []) or []:
-                    if getattr(t, 'is_connected', False):
-                        has_active_terminal = True
-                        break
-            elif hasattr(window, 'active_terminals') and self.connection in window.active_terminals:
-                terminal = window.active_terminals[self.connection]
-                # Check if the terminal is still valid and connected
-                if terminal and hasattr(terminal, 'is_connected'):
-                    has_active_terminal = terminal.is_connected
-            
-            # Update the connection's is_connected status
-            self.connection.is_connected = has_active_terminal
-            
-            # Log the status update for debugging
-            logger.debug(f"Updating status for {self.connection.nickname}: is_connected={has_active_terminal}")
-            
-            # Update the UI based on the connection status
-            if has_active_terminal:
-                self.status_icon.set_from_icon_name('network-idle-symbolic')
-                self.status_icon.set_tooltip_text(f'Connected to {getattr(self.connection, "hname", "") or self.connection.host}')
-                logger.debug(f"Set status icon to connected for {self.connection.nickname}")
-            else:
-                self.status_icon.set_from_icon_name('network-offline-symbolic')
-                self.status_icon.set_tooltip_text('Disconnected')
-                logger.debug(f"Set status icon to disconnected for {getattr(self.connection, 'nickname', 'connection')}")
-                
-            # Force a redraw to ensure the icon updates
-            self.status_icon.queue_draw()
-            
-        except Exception as e:
-            logger.error(f"Error updating status for {getattr(self.connection, 'nickname', 'connection')}: {e}")
-    
-    def update_display(self):
-        """Update the display with current connection data"""
-        # Update the labels with current connection data
-        if hasattr(self.connection, 'nickname') and hasattr(self, 'nickname_label'):
-            self.nickname_label.set_markup(f"<b>{self.connection.nickname}</b>")
-        
-        if hasattr(self.connection, 'username') and hasattr(self.connection, 'host') and hasattr(self, 'host_label'):
-            port_text = f":{self.connection.port}" if hasattr(self.connection, 'port') and self.connection.port != 22 else ""
-            self.host_label.set_text(f"{self.connection.username}@{self.connection.host}{port_text}")
-        # Refresh forwarding indicators if rules changed
-        self._update_forwarding_indicators()
-        
-        self.update_status()
-
-    def show_error(self, message):
-        """Show error message"""
-        dialog = Adw.MessageDialog(
-            transient_for=self,
-            heading='Error',
-            body=message,
-        )
-        dialog.add_response('ok', 'OK')
-        dialog.set_default_response('ok')
-        dialog.present()
 
 class WelcomePage(Gtk.Box):
     """Welcome page shown when no tabs are open"""
@@ -1114,7 +726,7 @@ class MainWindow(Adw.ApplicationWindow):
         self.connection_list.set_activate_on_single_click(False)  # Require double-click to activate
         
         # Set up drag and drop for reordering
-        self.setup_connection_list_dnd()
+        build_sidebar(self)
 
         # Right-click context menu to open multiple connections
         try:
@@ -1363,276 +975,6 @@ class MainWindow(Adw.ApplicationWindow):
         self.content_stack.set_visible_child_name("welcome")
         
         self._set_content_widget(self.content_stack)
-
-    def setup_connection_list_dnd(self):
-        """Set up drag and drop for connection list reordering"""
-        # Add drop target to the connection list
-        drop_target = Gtk.DropTarget.new(type=GObject.TYPE_PYOBJECT, actions=Gdk.DragAction.MOVE)
-        drop_target.connect('drop', self._on_connection_list_drop)
-        drop_target.connect('motion', self._on_connection_list_motion)
-        drop_target.connect('leave', self._on_connection_list_leave)
-        self.connection_list.add_controller(drop_target)
-        
-        # Store drop indicator state
-        self._drop_indicator_row = None
-        self._drop_indicator_position = None
-        
-        # Store ungrouped area state
-        self._ungrouped_area_row = None
-        self._ungrouped_area_visible = False
-    
-    def _on_connection_list_motion(self, target, x, y):
-        """Handle motion over the connection list for visual feedback"""
-        try:
-            # Clear previous indicator
-            self._clear_drop_indicator()
-            
-            # Show ungrouped area when dragging
-            self._show_ungrouped_area()
-            
-            # Find the row at the current position
-            row = self.connection_list.get_row_at_y(int(y))
-            if not row:
-                return Gdk.DragAction.MOVE
-            
-            # Check if this is the ungrouped area
-            if hasattr(row, 'ungrouped_area') and row.ungrouped_area:
-                # Show ungrouped area highlight
-                row.add_css_class('drag-over')
-                self._drop_indicator_row = row
-                self._drop_indicator_position = 'ungrouped'
-                return Gdk.DragAction.MOVE
-            
-            # Determine drop position (above/below the row)
-            row_y = row.get_allocation().y
-            row_height = row.get_allocation().height
-            relative_y = y - row_y
-            
-            if relative_y < row_height / 2:
-                position = 'above'
-            else:
-                position = 'below'
-            
-            # Show drop indicator
-            self._show_drop_indicator(row, position)
-            
-            return Gdk.DragAction.MOVE
-        except Exception as e:
-            logger.error(f"Error handling motion: {e}")
-            return Gdk.DragAction.MOVE
-    
-    def _on_connection_list_leave(self, target):
-        """Handle leaving the drop target area"""
-        self._clear_drop_indicator()
-        self._hide_ungrouped_area()
-        return True
-    
-    def _show_drop_indicator(self, row, position):
-        """Show visual drop indicator"""
-        try:
-            # Add CSS class to the row for visual feedback
-            if position == 'above':
-                row.add_css_class('drop-above')
-            else:
-                row.add_css_class('drop-below')
-            
-            self._drop_indicator_row = row
-            self._drop_indicator_position = position
-        except Exception as e:
-            logger.error(f"Error showing drop indicator: {e}")
-    
-    def _create_ungrouped_area(self):
-        """Create the ungrouped area row"""
-        if self._ungrouped_area_row:
-            return self._ungrouped_area_row
-            
-        ungrouped_area = Gtk.Box(orientation=Gtk.Orientation.VERTICAL, spacing=8)
-        ungrouped_area.add_css_class('ungrouped-area')
-        
-        # Add icon and label
-        icon = Gtk.Image.new_from_icon_name('folder-open-symbolic')
-        icon.set_pixel_size(24)
-        icon.add_css_class('dim-label')
-        
-        label = Gtk.Label(label=_("Drop connections here to ungroup them"))
-        label.add_css_class('dim-label')
-        label.add_css_class('caption')
-        
-        ungrouped_area.append(icon)
-        ungrouped_area.append(label)
-        
-        # Make it a list box row
-        ungrouped_row = Gtk.ListBoxRow()
-        ungrouped_row.set_child(ungrouped_area)
-        ungrouped_row.set_selectable(False)
-        ungrouped_row.set_activatable(False)
-        ungrouped_row.ungrouped_area = True
-        
-        self._ungrouped_area_row = ungrouped_row
-        return ungrouped_row
-    
-    def _show_ungrouped_area(self):
-        """Show the ungrouped area at the bottom of the list"""
-        try:
-            if self._ungrouped_area_visible:
-                return
-                
-            # Only show if there are groups
-            hierarchy = self.group_manager.get_group_hierarchy()
-            if not hierarchy:
-                return
-                
-            ungrouped_row = self._create_ungrouped_area()
-            self.connection_list.append(ungrouped_row)
-            self._ungrouped_area_visible = True
-            
-        except Exception as e:
-            logger.error(f"Error showing ungrouped area: {e}")
-    
-    def _hide_ungrouped_area(self):
-        """Hide the ungrouped area"""
-        try:
-            if not self._ungrouped_area_visible or not self._ungrouped_area_row:
-                return
-                
-            self.connection_list.remove(self._ungrouped_area_row)
-            self._ungrouped_area_visible = False
-            
-        except Exception as e:
-            logger.error(f"Error hiding ungrouped area: {e}")
-    
-    def _clear_drop_indicator(self):
-        """Clear visual drop indicator"""
-        try:
-            if self._drop_indicator_row:
-                self._drop_indicator_row.remove_css_class('drop-above')
-                self._drop_indicator_row.remove_css_class('drop-below')
-                self._drop_indicator_row.remove_css_class('drag-over')
-                self._drop_indicator_row = None
-                self._drop_indicator_position = None
-        except Exception as e:
-            logger.error(f"Error clearing drop indicator: {e}")
-    
-    def _on_connection_list_drop(self, target, value, x, y):
-        """Handle drops on the connection list"""
-        try:
-            # Clear drop indicator and hide ungrouped area
-            self._clear_drop_indicator()
-            self._hide_ungrouped_area()
-            
-            if not isinstance(value, dict):
-                return False
-            
-            drop_type = value.get('type')
-            if drop_type == 'connection':
-                connection_nickname = value.get('connection_nickname')
-                if connection_nickname:
-                    # Get current group of the connection
-                    current_group_id = self.group_manager.get_connection_group(connection_nickname)
-                    
-                    # Find the target row and position
-                    target_row = self.connection_list.get_row_at_y(int(y))
-                    if not target_row:
-                        # Dropping in empty space - move to ungrouped
-                        self.group_manager.move_connection(connection_nickname, None)
-                        self.rebuild_connection_list()
-                        return True
-                    
-                    # Check if dropping on ungrouped area
-                    if hasattr(target_row, 'ungrouped_area') and target_row.ungrouped_area:
-                        # Move to ungrouped
-                        self.group_manager.move_connection(connection_nickname, None)
-                        self.rebuild_connection_list()
-                        return True
-                    
-                    # Get drop position (above/below)
-                    row_y = target_row.get_allocation().y
-                    row_height = target_row.get_allocation().height
-                    relative_y = y - row_y
-                    position = 'above' if relative_y < row_height / 2 else 'below'
-                    
-                    # Determine if we're dropping on a group or connection
-                    if hasattr(target_row, 'group_id'):
-                        # Dropping on a group
-                        target_group_id = target_row.group_id
-                        if target_group_id != current_group_id:
-                            # Move to different group
-                            self.group_manager.move_connection(connection_nickname, target_group_id)
-                            self.rebuild_connection_list()
-                    else:
-                        # Dropping on a connection
-                        target_connection = getattr(target_row, 'connection', None)
-                        if target_connection:
-                            target_group_id = self.group_manager.get_connection_group(target_connection.nickname)
-                            if target_group_id != current_group_id:
-                                # Move to different group
-                                self.group_manager.move_connection(connection_nickname, target_group_id)
-                                self.rebuild_connection_list()
-                            else:
-                                # Reorder within the same group
-                                self.group_manager.reorder_connection_in_group(
-                                    connection_nickname, 
-                                    target_connection.nickname, 
-                                    position
-                                )
-                                self.rebuild_connection_list()
-                    
-                    return True
-            
-            elif drop_type == 'group':
-                group_id = value.get('group_id')
-                if group_id:
-                    # Handle group reordering
-                    target_row = self.connection_list.get_row_at_y(int(y))
-                    if target_row and hasattr(target_row, 'group_id'):
-                        target_group_id = target_row.group_id
-                        if target_group_id != group_id:
-                            self._move_group(group_id, target_group_id)
-                            self.rebuild_connection_list()
-                            return True
-            
-            return False
-        except Exception as e:
-            logger.error(f"Error handling drop: {e}")
-            return False
-    
-    def _get_target_group_at_position(self, x, y):
-        """Get the target group ID at the given position"""
-        try:
-            row = self.connection_list.get_row_at_y(int(y))
-            if row and hasattr(row, 'group_id'):
-                return row.group_id
-            elif row and hasattr(row, 'connection'):
-                # If dropping on a connection, get its group
-                connection = row.connection
-                return self.group_manager.get_connection_group(connection.nickname)
-            return None
-        except Exception:
-            return None
-    
-    def _move_group(self, group_id, target_parent_id):
-        """Move a group to a new parent"""
-        try:
-            if group_id not in self.group_manager.groups:
-                return
-            
-            group = self.group_manager.groups[group_id]
-            old_parent_id = group.get('parent_id')
-            
-            # Remove from old parent
-            if old_parent_id and old_parent_id in self.group_manager.groups:
-                if group_id in self.group_manager.groups[old_parent_id]['children']:
-                    self.group_manager.groups[old_parent_id]['children'].remove(group_id)
-            
-            # Add to new parent
-            group['parent_id'] = target_parent_id
-            if target_parent_id and target_parent_id in self.group_manager.groups:
-                if group_id not in self.group_manager.groups[target_parent_id]['children']:
-                    self.group_manager.groups[target_parent_id]['children'].append(group_id)
-            
-            self.group_manager._save_groups()
-        except Exception as e:
-            logger.error(f"Error moving group: {e}")
 
     def create_menu(self):
         """Create application menu"""
